pyproject.toml
[build-system]
requires = ["setuptools>=77.0"]
build-backend = "setuptools.build_meta"

[project]
name = "disco-torch"
version = "0.1.0"
description = "PyTorch port of DeepMind's Disco103 meta-learned RL update rule"
requires-python = ">=3.11"
license = "Apache-2.0"
dependencies = [
    "torch>=2.0",
    "numpy>=1.24",
]

[project.optional-dependencies]
dev = ["pytest", "jax", "dm-haiku", "rlax", "distrax"]

[tool.setuptools.packages.find]
include = ["disco_torch*"]
